Rearranged files and formatted them
