+++ src/modules/person/PersonTest.cpp
@@ -17,13 +17,13 @@
 #include "data/german/GermanFirstNamesFemales.h"
 #include "data/german/GermanFirstNamesMales.h"
 #include "data/german/GermanLastNames.h"
+#include "data/Hobbies.h"
 #include "data/indian/IndianFirstNames.h"
 #include "data/indian/IndianLastNames.h"
 #include "data/italian/ItalianFirstNamesFemales.h"
 #include "data/italian/ItalianFirstNamesMales.h"
 #include "data/italian/ItalianLastNames.h"
 #include "data/JobTitles.h"
-#include "data/hobby/Hobbies.h"
 #include "data/polish/PolishFirstNamesFemales.h"
 #include "data/polish/PolishFirstNamesMales.h"
 #include "data/polish/PolishLastNames.h"
@@ -323,6 +323,6 @@ TEST_F(PersonTest, shouldGenerateHobby)
 {
     const auto generatedHobby = Person::hobby();
 
-    ASSERT_TRUE(std::ranges::any_of(hobbies, [generatedHobby](const std::string& hobby)
-                                    { return hobby == generatedHobby; }));
+    ASSERT_TRUE(
+        std::ranges::any_of(hobbies, [generatedHobby](const std::string& hobby) { return hobby == generatedHobby; }));
 }
