+++ src/modules/person/data/Hobbies.h
@@ -1,0 +1,54 @@
+#pragma once
+
+#include <string>
+#include <vector>
+
+namespace faker
+{
+const std::vector<std::string> hobbies = {
+    "Playing guitar",
+    "Reading books",
+    "Hiking",
+    "Cooking",
+    "Painting",
+    "Photography",
+    "Swimming",
+    "Gaming",
+    "Dancing",
+    "Traveling",
+    "Bird watching",
+    "Coding",
+    "Singing",
+    "Collecting stamps",
+    "Fishing",
+    "Yoga",
+    "Sculpting",
+    "Mountain biking",
+    "Chess",
+    "Archery",
+    "Rock climbing",
+    "Skiing",
+    "Writing poetry",
+    "Meditation",
+    "Solving puzzles",
+    "Gardening",
+    "Playing basketball",
+    "Volunteering",
+    "Playing board games",
+    "Astrophotography",
+    "Camping",
+    "Model building",
+    "Astronomy",
+    "Watching movies",
+    "Stargazing",
+    "Playing tennis",
+    "Cooking ethnic cuisines",
+    "Woodworking",
+    "Baking",
+    "Calligraphy",
+    "Pottery",
+    "Karaoke",
+    "Scuba diving",
+    "Metal detecting",
+};
+}
